+++ src/components/home/StateOfJSIntro.js
@@ -10,10 +10,7 @@ const Intro = ({ date }) => (
   <div id="stateofjs-container">
     <div>
       <p>
-        <span
-          className="mega-octicon octicon-megaphone"
-          style={{ color: '#fa9e59' }}
-        />{' '}
+        <span className="mega-octicon octicon-megaphone icon-color" />{' '}
         <span style={{ fontSize: '1.5rem' }}>
           State of JavaScript 2017 Results
         </span>