+++ src/components/home/weekly-newsletter.js
@@ -3,13 +3,13 @@ import tinytime from 'tinytime'
 import styled from 'styled-components'
 import { useSelector } from 'react-redux'
 
+import { findProjectsByIds } from '../../selectors'
 import { ExternalLink, Section, Button } from '../core'
 import { getProjectId } from '../core/project'
 import { ChevronLeftIcon, ChevronRightIcon } from '../core/icons'
 import { useWeeklyNewsletter } from '../../api/hooks'
 import { Row, MainColumn, RightSideBar } from './layout'
 import ProjectList from '../project-list/ProjectTable'
-import { findProjectsByIds } from '../../selectors/project'
 import { SubscribeForm } from './subscribe-form'
 
 export const Weekly = () => {
@@ -122,7 +122,7 @@ const Story = styled.div`
   background-color: white;
   blockquote {
     padding: 0 1em;
-    border-left: 0.25em solid #fa9e59;
+    border-left: 0.25em solid var(--iconColor);
     margin-left: 0;
     color: var(--textSecondaryColor);
   }